+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+**Version v2 (master)**
+
+- Fix municipality type and canton imputation
 - Write out detailed information from MZ for PT trips
 - Fix education location
 - Add SP region and municipality type of household